requirements-test.txt
pytest==6.2.4
pytest-aiohttp==0.3.0
pytest-cov==2.12.1
pytest-timeout==1.4.2